# sample_project/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Swaggelize API",
        "description": "API automatic generator using Swagger API and Swaggelize",
        "contact": {
            "name": "Hasininjato Rojovao"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api"
        },
        {
            "url": "http://localhost:3000/api"
        }
    ],
    "components": {
...
        "schemas": {
...
            "userItem": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int32",
                        "description": "Id of the user"
                    },
                    "fullname": {
                        "type": "string",
                        "description": "User fullname"
                    },
                    "email": {
                        "type": "string",
                        "description": "Email user"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time",
                        "description": "Date when the record was created"
                    },
                    "updatedAt": {
                        "type": "string",
                        "format": "date-time",
                        "description": "Date when the record was last updated"
                    }
                }
            },
...
            "Response400Schema": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string",
                        "description": "Name of the error",
                        "example": "ValidationError",
                        "enum": [
                            "ValidationError",
                            "UniqueConstraintError"
                        ]
                    },
                    "errors": {
                        "type": "array",
                        "items": {
                            "type": "object",
                            "properties": {
                                "field": {
                                    "type": "string",
                                    "description": "Name of the field"
                                },
                                "message": {
                                    "type": "string",
                                    "description": "Message on validation"
                                }
                            }
                        }
                    }
                }
            },
            "Response409Schema": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string",
                        "description": "Name of the error",
                        "example": "ValidationError",
                        "enum": [
                            "ValidationError",
                            "UniqueConstraintError"
                        ]
                    },
                    "errors": {
                        "type": "array",
                        "items": {
                            "type": "object",
                            "properties": {
                                "field": {
                                    "type": "string",
                                    "description": "Name of the field"
                                },
                                "message": {
                                    "type": "string",
                                    "description": "Message on validation"
                                }
                            }
                        }
                    }
                }
            },
...
            "CustomLogin": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int32",
                        "description": "Id of the user"
                    },
                    "fullname": {
                        "type": "string",
                        "description": "User fullname"
                    },
                    "email": {
                        "type": "string",
                        "description": "Email user"
                    },
                    "access_token": {
                        "type": "string",
                        "description": "Access token"
                    },
                    "refresh_token": {
                        "type": "string",
                        "description": "Refresh token",
                        "example": "fkasjd;ur  rqfj asudf"
                    },
                    "array_tokens": {
                        "type": "array",
                        "description": "",
                        "items": {
                            "type": "object",
                            "properties": {
                                "access_token": {
                                    "type": "string",
                                    "description": "Access token"
                                },
                                "refresh_token": {
                                    "type": "string",
                                    "description": "Refresh token"
                                }
                            }
                        }
                    },
                    "object_tokens": {
                        "type": "object",
                        "description": "",
                        "properties": {
                            "access_token": {
                                "type": "string",
                                "description": "Access token"
                            },
                            "refresh_token": {
                                "type": "string",
                                "description": "Refresh token",
                                "example": "ejgh8ta582.afkrioakljflkjsdfk"
                            }
                        }
                    }
                }
            },
            "CustomValidateToken": {
                "type": "object",
                "properties": {
                    "message": {
                        "type": "string",
                        "description": "Validity of the token",
                        "example": "The token is still valid"
                    }
                }
            },
...
        }
    },
    "paths": {
...
        "/auth/login": {
            "post": {
                "summary": "Login an user",
                "description": "Login an user",
                "tags": [
                    "Authentication"
                ],
                "responses": {
                    "200": {
                        "description": "Successfully logged in",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CustomLogin"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request (VALIDATION_ERROR)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Response400Schema"
                                },
                                "example": {
                                    "name": "ValidationError",
                                    "errors": []
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid credentials"
                    },
                    "403": {
                        "description": "Invalid password or email. Please try again"
                    },
                    "409": {
                        "description": "Unique constraint errors",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Response409Schema"
                                },
                                "example": {
                                    "name": "UniqueConstraintError",
                                    "errors": []
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                },
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/userItem"
                            }
                        }
                    }
                }
            }
        },
        "/validate/token": {
            "get": {
                "summary": "Validate token sent in parameter",
                "description": "Validate token sent in parameter",
                "tags": [
                    "Token"
                ],
                "responses": {
                    "200": {
                        "description": "Token is valid",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CustomValidateToken"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid credentials"
                    },
                    "403": {
                        "description": "Invalid password or email. Please try again"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                },
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/userItem"
                            }
                        }
                    }
                }
            }
        },
...
    }
}